+++ src/index.ts
@@ -75,4 +75,4 @@ export * from './app/split-view/multi/data/terra-multi-split-view.interface';
 export * from './app/dynamic-module-loader/data/terra-split-view-component.interface';
 export * from './app/dynamic-component-loader/terra-dynamic-component-loader.component';
 export * from './app/no-result/terra-no-result-notice.component';
-export * from './app/wysiwyg-editor/terra-wysiwyg-editor.component';
+export * from './app/notes-editor/terra-notes-editor.component';